Move hdri-shading to new system
