+++ src/hdri-shading.html
@@ -8,12 +8,13 @@
 		<title>Initial Document</title>
 	</head>
 	<body>
-		<script type="module" src="./js/hdri-shading.js"></script>
+		<script type="module" src="./js/views/hdri-shading.js"></script>
+		<script src="./js/inline.js"></script>
 		<main id="renderer_target">
 			<nav>
 				<label pbr1-onlyshowifmultivalued="environment_name" for="environment_name" title="environment_name">Environment</label>
 				<select 
-					onchange="updateScene({'environment_index':this.value},PBR1_FALLBACK.current);"
+					onchange="notifyChange({'environment_index':this.value});"
 					pbr1-onlyshowifmultivalued="environment_name"
 					pbr1-optionsource="environment_name"
 					pbr1-selectedsource="environment_index"
@@ -22,7 +23,7 @@
 				>
 				</select>
 
-				<label ondblclick="updateScene({'environment_exposure':PBR1_SCENECONFIG.default.environment_exposure},PBR1_FALLBACK.current);"; for="environment_exposure" title="environment_exposure">Exposure</label>
+				<label id="environment_exposure_label" for="environment_exposure" title="environment_exposure">Exposure</label>
 				<input 
 					type="range"
 					min="-16"
@@ -30,23 +31,18 @@
 					value="0"
 					id="environment_exposure"
 					step="0.25"
-					ondblclick="updateScene({'environment_exposure':PBR1_SCENECONFIG.default.environment_exposure},PBR1_FALLBACK.current);";
-					oninput="updateScene({'environment_exposure':this.value},PBR1_FALLBACK.current);",
-					onload="updateScene({'environment_exposure':this.value},PBR1_FALLBACK.current);"
+					oninput="notifyChange({'environment_exposure':this.value});"
+					ondblclick="notifyChange({'environment_exposure':null},true);"
 				>
 				<label for="spheres_enable" title="spheres_enable">Show Spheres</label>
 				<input 
 					checked="checked"
 					type="checkbox"
 					id="spheres_enable"
-					onchange="updateScene({'spheres_enable':this.checked ? 1 : 0},PBR1_FALLBACK.current);"
+					onchange="notifyChange({'spheres_enable':this.checked ? 1 : 0});"
 				>
 			</nav>
 			<a class="watermark watermark-off" href="https://pbr.one">🔮 PBR.ONE</a>
 		</main>
-		<!--<script src="./lib/three.r142.min.js"></script>
-		<script src="./lib/OrbitControls.r142.js"></script>
-		<script src="./lib/RGBELoader.r142.js"></script>
-		<script src="./lib/EXRLoader.r142.js"></script>-->
 	</body>
 </html>
