+++ src/js/common/scenestate.js
